I changed all of the number variables to a double variable, because I would only get an int as a result, which is why I had a bug

+++ src/View/Addition.java
@@ -3,7 +3,7 @@ package View;
 public class Addition implements OperationsLogic{
 
     @Override
-    public double operation(int number1, int number2)
+    public double operation(double number1, double number2)
     {
         return number1 + number2;
     }
