+++ src/View/Reciprocal.java
@@ -2,7 +2,7 @@ package View;
 
 public class Reciprocal implements OperationsLogic{
     @Override
-    public double operation(int number1, int number2) {
+    public double operation(double number1, double number2) {
         return 1 / number1;
     }
